requirements.txt
Django==5.0.1
mongoengine==0.27.0
python-dotenv==1.0.0